refactor: rename Parameters folder to parameters
