+++ source/controllers/include/controllers/impedance/Dissipative.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "controllers/impedance/Impedance.hpp"
-#include "state_representation/Parameters/Parameter.hpp"
+#include "state_representation/parameters/Parameter.hpp"
 #include "state_representation/Space/Cartesian/CartesianState.hpp"
 #include <eigen3/Eigen/Core>
 #include <eigen3/Eigen/Dense>
