+++ source/robot_model/include/robot_model/Model.hpp
@@ -8,8 +8,8 @@
 #include <pinocchio/multibody/data.hpp>
 #include <pinocchio/algorithm/crba.hpp>
 #include <pinocchio/algorithm/rnea.hpp>
-#include <state_representation/Parameters/Parameter.hpp>
-#include <state_representation/Parameters/ParameterInterface.hpp>
+#include <state_representation/parameters/Parameter.hpp>
+#include <state_representation/parameters/ParameterInterface.hpp>
 #include <state_representation/Robot/Jacobian.hpp>
 #include <state_representation/Robot/JointState.hpp>
 #include <state_representation/Space/Cartesian/CartesianState.hpp>
