+++ source/state_representation/include/state_representation/parameters/Parameter.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "state_representation/Parameters/ParameterInterface.hpp"
+#include "state_representation/parameters/ParameterInterface.hpp"
 
 namespace state_representation {
 template <typename T>
