+++ source/state_representation/src/parameters/Event.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/Parameters/Event.hpp"
+#include "state_representation/parameters/Event.hpp"
 
 namespace state_representation {
 Event::Event(const std::string& name) : Predicate(name), previous_predicate_value_(false) {}
