+++ source/state_representation/src/parameters/Parameter.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/Parameters/Parameter.hpp"
+#include "state_representation/parameters/Parameter.hpp"
 #include "state_representation/geometry/Ellipsoid.hpp"
 #include "state_representation/Robot/JointPositions.hpp"
 #include "state_representation/Space/Cartesian/CartesianPose.hpp"
