+++ source/state_representation/src/parameters/Predicate.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/Parameters/Predicate.hpp"
+#include "state_representation/parameters/Predicate.hpp"
 
 namespace state_representation {
 Predicate::Predicate(const std::string& name) : Parameter<bool>(name, false) {}
